f.2.3 clean up and more coop rule changes (#166)
* f.2.3 clean up and more coop rule changes
f.2.3 does not work with realms, but we have to determine an actual policy for that before it's added to the rules. I think f.2.3 should probably be struck entirely and the run should start when the first person joins b/c the logs would also tell us that but that is tbd.

* entangle rule

* passive voice